% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbmovMF.R
\name{dbmovMF}
\alias{dbmovMF}
\title{Diagonal Block Mixture of von Mises-Fisher Dsitributions}
\usage{
dbmovMF(X, k, control = list(), ...)
}
\arguments{
\item{X}{A sparse or dense data matrix, where rows denote objets lying on a unit-hypersphere and column denote features. Supported format include Matrix, dgCMatrix, dgTMatrix, etc.}

\item{k}{The number of mixture compoenents (or clusters).}

\item{control}{A list if additional parameters, see Details.}

\item{...}{A list of additional parameters (overriding those specified in control).}
}
\value{
returns an object of class dbmovMF including the following attributes:
\itemize{
\item rowcluster: the partition of rows into k clusters.
\item colcluster: the partition of columns into k clusters.
\item kappa_: a vector of concentration parameters.
\item alpha: a vector of cluster proportions.
\item alpha: a vector of cluster proportions.
\item ll: the values of the complete data log likelihood over iterations.
\item Iter: the total number of iteration needed for convergence.
}
}
\description{
Provides a simple function to bluid and fit a block mixture of von Mises-Fisher for data co-clustering.
}
\details{
In addition to the parameters described above, the function dbmovMF supports other useful parameters including:
\itemize{
\item max_iter: the maximum number of iteration
\item max_st_iter: the maximum number of stochatic iterations. Only useful when performing learning using either SAEMb or CAEMb
\item n_init: number of time the algorithm will be run with different initializations. The final results correspond to the bust run in terms of log likelihood.
\item tol: tolerance relatve to the log likelihood to declar convergence.
\item equal_prop: if true then then all mixture proportion will be consider to be equal, and the mixture propotions parmameters alpha will be ignored.
\item fit_algo: learning algorithm to use. Supported algorithms include: EMb, CEMb, SEMb, SAEMb and CAEMb.
\item equal_kappa: if true then all concentration parameters kappa of the differrent mixture components will be equal
\item kappa_: The value of the concentration parameter. This parameter is only effective when euqal_kappa = TRUE.
\item row_init: a partition of rows into k cluster to be used for initialization. Could be a vector if n_init = 1 of a matrix of size (n_init x n_rows) if n_init>1. Default = NULL.
\item col_init: same as row_init but for columns.
}
}
